Refactor auth folder + Adding local auth guard class
